rename SearchContainer and add in the search icon from the prev version
